add file zod type for datas
