+++ src/features/characters/PageCharacters.tsx
@@ -2,54 +2,20 @@ import React from 'react';
 
 import { Button, Divider, Heading, Stack, Text } from '@chakra-ui/react';
 import { useQuery } from '@tanstack/react-query';
-import { N } from 'vitest/dist/chunks/reporters.nr4dxCkA.js';
-import { z } from 'zod';
 
 import { ErrorPage } from '@/components/ErrorPage';
 import { LoaderFull } from '@/components/LoaderFull';
 
 import { AppLayoutPage } from '../app/AppLayoutPage';
 import CardCharacter from './CardCharacter';
+import { CharactersAPI } from './CharacterZod';
 
 export default function PageCharacters() {
-  const characterSchema = z.object({
-    info: z.object({
-      count: z.number(),
-      pages: z.number(),
-      next: z.string(),
-      prev: z.string(),
-    }),
-    results: z.array(
-      z.object({
-        id: z.number(),
-        name: z.string(),
-        status: z.string(),
-        species: z.string(),
-        type: z.string(),
-        gender: z.string(),
-        origin: z.object({
-          name: z.string(),
-          url: z.string(),
-        }),
-        location: z.object({
-          name: z.string(),
-          url: z.string(),
-        }),
-        image: z.string(),
-        episode: z.array(z.string()),
-        url: z.string(),
-        created: z.string(),
-      })
-    ),
-  });
-
-  type CharacterAPI = z.infer<typeof characterSchema>;
-
   const [url, setUrl] = React.useState(
     'https://rickandmortyapi.com/api/character'
   );
 
-  const { data, isLoading, isError } = useQuery<CharacterAPI>(
+  const { data, isLoading, isError } = useQuery<CharactersAPI>(
     ['searchAllCharacters', url],
     async () => {
       const response = await fetch(url);
